Part3/pyproject.toml
[tool.poetry]
name = "part1"
version = "0.1.0"
description = ""
authors = ["SwanyCastle <seunghwan712@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.2"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
